debian: mask mmsd-tng when mmsd4ofono is installed
if mmsd-tng is installed alongside our mmsd implementation, it will try to steal the bus which blocks our mmsd from listening on user session.
the more "correct" way of solving this is to add a conflicts in debian/control to make sure mmsd-tng won't get installed at all if our mmsd is installed,
but debian meta packages add mmsd-tng so depends and recommends, so they will break. you win some you lose some. we can come back to this later if it causes any issues

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>