ofono_push_notification: export received messages with empty string a smil if they don't have any smil
some mms PDUs might not include smil, in this case, mmsd used to save the PDU and attachments but never export onto dbus. instead, just pass an empty string as smil. not a perfect solution but it is what it is
tested and confirmed to be an issue with messages coming from Nokia 225 4g 2024 dumbphone running T-Mobile

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>