main: give up on context activation if internet and mms APNs are the same
in case the APN for data and MMS ends up the sane, MMSD will try to activate the context and get stuck in an infinite loop in an attempt to steal the context from internet.
this goes on until ofono2mm deactivates the context, and then MMS takes over the APN which causes ofono2mm to never get the context back.
the reported solution was to "turn airplane mode on and off then toggle mobile data" which is not at all ideal.
a much more correct solution would be to support mms+internet APNs, but we don't, so lets at least not break data for users

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>